Add files via upload
These MatLab files of code finds the correlation, Euclidean distance,   between two gene expression microarray data. It finds one gene with all genes and we need to set a threshold then we can say that these two genes are correlated as well as they are in Euclidean space. I proposed an algorithem i.e. GeCon which have both the property and also consider local as well as global expression pattern.